% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIproj-package.R
\docType{package}
\name{SCIproj-package}
\alias{SCIproj}
\alias{SCIproj-package}
\title{SCIproj package}
\description{
Generic template for research projects structured as R packages.
This package works as a template for new research projects,
under the idea of having everything (data, R scripts, functions and manuscripts)
contained in the same package to promote reproducibility.
}
\details{
For detailed steps on how to use the package see \url{https://github.com/saskiaotto/SCIproj/blob/master/README.md}.
}
\author{
Saskia A. Otto
}
\keyword{internal}
