% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proj.R
\name{create_proj}
\alias{create_proj}
\title{Create new data analysis project as R package}
\usage{
create_proj(
  name,
  data_raw = TRUE,
  makefile = FALSE,
  testthat = FALSE,
  use_pipe = TRUE,
  add_license = NULL,
  license_holder = "Your name",
  create_github_repo = FALSE,
  private_repo = TRUE,
  ci = "none",
  open_proj = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{name}{Character. Name of the new project. Could be a path, e.g. \code{"~/myRcode/newproj"}. A new folder will be created with that name.}

\item{data_raw}{Logical. If TRUE, adds a 'data-raw' folder for all the raw data files, including a 'clean_data.R' script file for generating
clean datasets from the raw files that are then saved in the 'data' folder.}

\item{makefile}{Logical. If TRUE, adds a template \code{makefile.R} file to the project.}

\item{testthat}{Logical. Add testthat infrastructure?}

\item{use_pipe}{Logical. Use magrittr's pipe in your package?}

\item{add_license}{Character. Use magrittr's pipe in your package?}

\item{license_holder}{Character. If you want to use a license than replace 'Your name' with your own name.}

\item{create_github_repo}{Logical. Create GitHub repository? Note this requires some working infrastructure like \code{git} and a \code{GITHUB_PAT}. See instructions here \url{https://usethis.r-lib.org/articles/articles/usethis-setup.html}.}

\item{private_repo}{Logical. Default is TRUE.}

\item{ci}{Logical. Use continuous integration in your GitHub repository? Current options are "none" (default), "travis" (uses Travis-CI), "circle" (uses Circle-CI), "appveyor" (uses AppVeyor), or "gh-actions" (uses GitHub Actions).}

\item{open_proj}{Logical. If TRUE (the default) will open the newly created Rstudio project in a new session.}

\item{verbose}{Logical. Print verbose output in the console while creating the new project? Default is FALSE.}
}
\value{
A new directory with a slightly modified R package structure.
}
\description{
Create all the scaffolding for a new project in a new directory. The scaffolding includes a README file, different folders to hold raw data, analyses, etc, and optionally also \code{testthat} infrastructure. Also, optionally, set a private or public GitHub repo with continuous integration (Travis-CI, GitHub Actions...).
}
\details{
If using github = TRUE, you will be asked if you want to commit some files. Reply positively to continue.
}
\examples{
\dontrun{
library("SCIproj")
create_proj("myproject")
create_proj("myproject", github = TRUE, private.repo = TRUE)
}
}
